Complete server

+++ pkg/responses/responcses.go
@@ -1,7 +1,7 @@
 package responses
 
 type TimerResponse struct {
-	SecondsRemaining int `json:"seconds-remaining"`
+	SecondsRemaining int64 `json:"seconds-remaining"`
 }
 
 type BitlyResponse struct {
